# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: StatusPage API
  description: API for managing service statuses and incidents
  version: 2.15.0
  contact:
    name: API Support
servers:
  - url: /
    description: Relative paths (for testing)
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Log in to the system
      description: |
        Authenticates user and sets HTTP-only cookies for tokens.

        **Cookies set:**
        - `access_token` - JWT access token (HttpOnly, Secure, SameSite=Lax)
        - `refresh_token` - JWT refresh token (HttpOnly, Secure, SameSite=Strict, Path=/api/v1/auth)
        - `csrf_token` - CSRF token for subsequent requests (readable by JavaScript)
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful. Tokens are set via Set-Cookie headers.
          headers:
            Set-Cookie:
              description: Authentication cookies (access_token, refresh_token, csrf_token)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Refresh tokens
      description: |
        Refreshes authentication tokens using the refresh_token cookie.
        New tokens are set via Set-Cookie headers.

        **For cookie-based auth:** No request body needed, refresh_token is read from cookie.
        **For API clients:** Can pass refresh_token in request body.
      operationId: refreshTokens
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Tokens refreshed. New tokens are set via Set-Cookie headers.
          headers:
            Set-Cookie:
              description: New authentication cookies
              schema:
                type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: Log out from the system
      description: |
        Invalidates the refresh token and clears all authentication cookies.

        **For cookie-based auth:** No request body needed, refresh_token is read from cookie.
        **For API clients:** Can pass refresh_token in request body.
      operationId: logout
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Logout successful. Cookies are cleared via Set-Cookie headers.
          headers:
            Set-Cookie:
              description: Cleared authentication cookies (Max-Age=0)
              schema:
                type: string
  /api/v1/auth/forgot-password:
    post:
      tags: [auth]
      summary: Request password reset
      description: |
        Sends a password reset link to the user's email.
        Always returns 200 to prevent email enumeration (except when email is not configured).
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Reset link sent (or silently ignored if email not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Email not configured or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/reset-password:
    post:
      tags: [auth]
      summary: Reset password using token
      description: Sets a new password using a valid reset token. Token is single-use and expires after 1 hour.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Password reset successful
        '400':
          description: Invalid or expired token, or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/me:
    get:
      tags: [auth]
      summary: Get current user
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags: [auth]
      summary: Update current user profile
      operationId: updateProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/me/password:
    put:
      tags: [auth]
      summary: Change current user password
      description: |
        Changes the authenticated user's password. Requires the current password.
        On success, all refresh tokens are invalidated and auth cookies are cleared.
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed. Auth cookies are cleared.
          headers:
            Set-Cookie:
              description: Cleared authentication cookies (Max-Age=0)
              schema:
                type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/users:
    get:
      tags: [users]
      summary: List users
      description: Admin-only. Returns paginated list of users with optional role filter.
      operationId: listUsers
      security:
        - BearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [user, operator, admin]
          description: Filter by user role
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [users]
      summary: Create a user
      description: |
        Admin-only. Creates a new user with specified role.
        Sets must_change_password=true and creates a default email channel.
      operationId: adminCreateUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error or invalid role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/users/{id}:
    get:
      tags: [users]
      summary: Get user details
      description: Admin-only. Returns user details by ID.
      operationId: getUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [users]
      summary: Update a user
      description: |
        Admin-only. Update user role, active status, or profile fields.
        Cannot modify your own account (returns 409).
        On deactivation, all refresh tokens are invalidated.
      operationId: adminUpdateUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error or invalid role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Cannot modify own account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/users/{id}/reset-password:
    post:
      tags: [users]
      summary: Reset user password
      description: |
        Admin-only. Sets a new password for the user.
        Sets must_change_password=true and invalidates all refresh tokens.
        Cannot reset your own password via this endpoint (returns 409).
      operationId: adminResetPassword
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminResetPasswordRequest'
      responses:
        '204':
          description: Password reset successful
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Cannot reset own password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        For API clients (curl, Postman, integrations).
        Pass the access token in the Authorization header.
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        For browser-based clients.
        Cookies are automatically set by /auth/login and sent by the browser.
        State-changing requests (POST, PUT, PATCH, DELETE) require X-CSRF-Token header
        matching the csrf_token cookie value.
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                  details:
                    oneOf:
                      - type: array
                        items:
                          type: object
                          properties:
                            field:
                              type: string
                            message:
                              type: string
                          required: [field, message]
                      - type: string
                    description: Validation error details. Array of field errors or error string.
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
...
    ConflictError:
      description: Conflict
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
  schemas:
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        first_name:
          type: string
        last_name:
          type: string
      required: [email, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]
...
    UserResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/User'
    LoginResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
...
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
...