# internal/identity/handler.go
package identity

import (
	"crypto/rand"
	"encoding/hex"
	"encoding/json"
	"errors"
	"net/http"
	"strconv"
	"time"

	"github.com/bissquit/incident-garden/internal/domain"
	"github.com/bissquit/incident-garden/internal/pkg/ctxlog"
	"github.com/bissquit/incident-garden/internal/pkg/httputil"
	"github.com/go-chi/chi/v5"
	"github.com/go-playground/validator/v10"
)

var errorMappings = []httputil.ErrorMapping{
	{Error: ErrUserNotFound, Status: http.StatusNotFound},
	{Error: ErrEmailExists, Status: http.StatusConflict},
	{Error: ErrInvalidCredentials, Status: http.StatusUnauthorized},
	{Error: ErrInvalidToken, Status: http.StatusUnauthorized},
	{Error: ErrAccountDeactivated, Status: http.StatusForbidden},
	{Error: ErrWrongPassword, Status: http.StatusBadRequest, Message: "wrong current password"},
	{Error: ErrEmailNotConfigured, Status: http.StatusBadRequest, Message: "email is not configured, contact your administrator"},
	{Error: ErrInvalidResetToken, Status: http.StatusBadRequest, Message: "invalid or expired reset token"},
	{Error: ErrCannotModifySelf, Status: http.StatusConflict, Message: "cannot modify your own account"},
	{Error: ErrInvalidRole, Status: http.StatusBadRequest, Message: "invalid role"},
}

// Pagination defaults for user listing.
const (
	DefaultUsersLimit = 20
	MaxUsersLimit     = 100
)

// CookieSettings contains settings for authentication cookies.
type CookieSettings struct {
	Secure               bool
	Domain               string
	AccessTokenDuration  time.Duration
	RefreshTokenDuration time.Duration
}

// Handler handles HTTP requests for the identity module.
type Handler struct {
	service        *Service
	validator      *validator.Validate
	cookieSettings CookieSettings
}

// NewHandler creates a new identity handler.
func NewHandler(service *Service, cookieSettings CookieSettings) *Handler {
	return &Handler{
		service:        service,
		validator:      validator.New(),
		cookieSettings: cookieSettings,
	}
}

// RegisterRoutes registers identity routes.
func (h *Handler) RegisterRoutes(r chi.Router) {
	r.Route("/auth", func(r chi.Router) {
		r.Post("/register", h.Register)
		r.Post("/login", h.Login)
		r.Post("/refresh", h.Refresh)
		r.Post("/logout", h.Logout)
		r.Post("/forgot-password", h.ForgotPassword)
		r.Post("/reset-password", h.ResetPassword)
	})
}

// RegisterProtectedRoutes registers routes that require authentication.
func (h *Handler) RegisterProtectedRoutes(r chi.Router) {
	r.Get("/me", h.Me)
	r.Patch("/me", h.UpdateProfile)
	r.Put("/me/password", h.ChangePassword)
}

// RegisterRequest represents registration request body.
type RegisterRequest struct {
	Email     string `json:"email" validate:"required,email"`
	Password  string `json:"password" validate:"required,min=8"`
	FirstName string `json:"first_name"`
	LastName  string `json:"last_name"`
}

// Register handles POST /auth/register.
func (h *Handler) Register(w http.ResponseWriter, r *http.Request) {
	var req RegisterRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		httputil.Error(w, http.StatusBadRequest, "invalid json")
		return
	}

	if err := h.validator.Struct(req); err != nil {
		httputil.ValidationError(w, err)
		return
	}

	user, err := h.service.Register(r.Context(), RegisterInput(req))
	if err != nil {
		httputil.HandleError(r.Context(), w, err, errorMappings)
		return
	}

	httputil.Success(w, http.StatusCreated, user)
}

// LoginRequest represents login request body.
type LoginRequest struct {
	Email    string `json:"email" validate:"required,email"`
	Password string `json:"password" validate:"required"`
}

// LoginResponse represents login response.
type LoginResponse struct {
	User *domain.User `json:"user"`
}

// Login handles POST /auth/login.
func (h *Handler) Login(w http.ResponseWriter, r *http.Request) {
	var req LoginRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		httputil.Error(w, http.StatusBadRequest, "invalid json")
		return
	}

	if err := h.validator.Struct(req); err != nil {
		httputil.ValidationError(w, err)
		return
	}

	user, tokens, err := h.service.Login(r.Context(), LoginInput(req))
	if err != nil {
		httputil.HandleError(r.Context(), w, err, errorMappings)
		return
	}

	h.setAuthCookies(w, tokens)

	httputil.Success(w, http.StatusOK, LoginResponse{
		User: user,
	})
}

// Refresh handles POST /auth/refresh.
// Reads refresh_token from cookie, issues new tokens.
func (h *Handler) Refresh(w http.ResponseWriter, r *http.Request) {
	refreshToken := h.getRefreshTokenFromRequest(r)
	if refreshToken == "" {
		httputil.Error(w, http.StatusBadRequest, "missing refresh token")
		return
	}

	tokens, err := h.service.RefreshTokens(r.Context(), refreshToken)
	if err != nil {
		httputil.HandleError(r.Context(), w, err, errorMappings)
		return
	}

	h.setAuthCookies(w, tokens)

	w.WriteHeader(http.StatusNoContent)
}

// Logout handles POST /auth/logout.
// Reads refresh_token from cookie, invalidates it, clears all auth cookies.
func (h *Handler) Logout(w http.ResponseWriter, r *http.Request) {
	refreshToken := h.getRefreshTokenFromRequest(r)
	if refreshToken != "" {
		if err := h.service.Logout(r.Context(), refreshToken); err != nil {
			ctxlog.FromContext(r.Context()).Warn("logout error", "error", err)
		}
	}

	h.clearAuthCookies(w)

	w.WriteHeader(http.StatusNoContent)
}

// Me handles GET /me.
func (h *Handler) Me(w http.ResponseWriter, r *http.Request) {
	userID := httputil.GetUserID(r.Context())
	if userID == "" {
		httputil.Error(w, http.StatusUnauthorized, "unauthorized")
		return
	}

	user, err := h.service.GetUserByID(r.Context(), userID)
	if err != nil {
		httputil.HandleError(r.Context(), w, err, errorMappings)
		return
	}

	httputil.Success(w, http.StatusOK, user)
}

// ChangePasswordRequest represents password change request body.
type ChangePasswordRequest struct {
	CurrentPassword string `json:"current_password" validate:"required"`
	NewPassword     string `json:"new_password" validate:"required,min=8"`
}

// ChangePassword handles PUT /me/password.
func (h *Handler) ChangePassword(w http.ResponseWriter, r *http.Request) {
	userID := httputil.GetUserID(r.Context())

	var req ChangePasswordRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		httputil.Error(w, http.StatusBadRequest, "invalid json")
		return
	}

	if err := h.validator.Struct(req); err != nil {
		httputil.ValidationError(w, err)
		return
	}

	err := h.service.ChangePassword(r.Context(), ChangePasswordInput{
		UserID:          userID,
		CurrentPassword: req.CurrentPassword,
		NewPassword:     req.NewPassword,
	})
	if err != nil {
		httputil.HandleError(r.Context(), w, err, errorMappings)
		return
	}

	h.clearAuthCookies(w)
	w.WriteHeader(http.StatusNoContent)
}

// UpdateProfileRequest represents profile update request body.
type UpdateProfileRequest struct {
	FirstName *string `json:"first_name" validate:"omitempty,max=100"`
	LastName  *string `json:"last_name" validate:"omitempty,max=100"`
}

// UpdateProfile handles PATCH /me.
func (h *Handler) UpdateProfile(w http.ResponseWriter, r *http.Request) {
	userID := httputil.GetUserID(r.Context())

	var req UpdateProfileRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		httputil.Error(w, http.StatusBadRequest, "invalid json")
		return
	}

	if err := h.validator.Struct(req); err != nil {
		httputil.ValidationError(w, err)
		return
	}

	user, err := h.service.UpdateProfile(r.Context(), UpdateProfileInput{
		UserID:    userID,
		FirstName: req.FirstName,
		LastName:  req.LastName,
	})
	if err != nil {
		httputil.HandleError(r.Context(), w, err, errorMappings)
		return
	}

	httputil.Success(w, http.StatusOK, user)
}

// ForgotPasswordRequest represents forgot password request body.
type ForgotPasswordRequest struct {
	Email string `json:"email" validate:"required,email"`
}

// ForgotPassword handles POST /auth/forgot-password.
// Always returns 200 to prevent email enumeration, except when email is not configured (400).
func (h *Handler) ForgotPassword(w http.ResponseWriter, r *http.Request) {
	var req ForgotPasswordRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		httputil.Error(w, http.StatusBadRequest, "invalid json")
		return
	}

	if err := h.validator.Struct(req); err != nil {
		httputil.ValidationError(w, err)
		return
	}

	err := h.service.ForgotPassword(r.Context(), req.Email)
	if err != nil {
		if errors.Is(err, ErrEmailNotConfigured) {
			httputil.HandleError(r.Context(), w, err, errorMappings)
			return
		}
		// Log but don't expose internal errors
		ctxlog.FromContext(r.Context()).Error("forgot password error", "error", err)
	}

	httputil.Success(w, http.StatusOK, map[string]string{
		"message": "if the email exists, a reset link has been sent",
	})
}

// ResetPasswordRequest represents reset password request body.
type ResetPasswordRequest struct {
	Token       string `json:"token" validate:"required"`
	NewPassword string `json:"new_password" validate:"required,min=8"`
}

// ResetPassword handles POST /auth/reset-password.
func (h *Handler) ResetPassword(w http.ResponseWriter, r *http.Request) {
	var req ResetPasswordRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		httputil.Error(w, http.StatusBadRequest, "invalid json")
		return
	}

	if err := h.validator.Struct(req); err != nil {
		httputil.ValidationError(w, err)
		return
	}

	err := h.service.ResetPassword(r.Context(), ResetPasswordInput(req))
	if err != nil {
		httputil.HandleError(r.Context(), w, err, errorMappings)
		return
	}

	w.WriteHeader(http.StatusNoContent)
}

// RegisterAdminRoutes registers admin-only user management routes.
func (h *Handler) RegisterAdminRoutes(r chi.Router) {
	r.Route("/users", func(r chi.Router) {
		r.Get("/", h.ListUsers)
		r.Post("/", h.AdminCreateUser)
		r.Get("/{id}", h.GetUser)
		r.Patch("/{id}", h.AdminUpdateUser)
		r.Post("/{id}/reset-password", h.AdminResetPassword)
	})
}

// ListUsers handles GET /users.
func (h *Handler) ListUsers(w http.ResponseWriter, r *http.Request) {
	filter := UserFilter{
		Limit: DefaultUsersLimit,
	}

	if roleParam := r.URL.Query().Get("role"); roleParam != "" {
		role := domain.Role(roleParam)
		if !role.IsValid() {
			httputil.Error(w, http.StatusBadRequest, "invalid role filter")
			return
		}
		filter.Role = &role
	}

	if l := r.URL.Query().Get("limit"); l != "" {
		parsed, err := strconv.Atoi(l)
		if err != nil || parsed < 1 {
			httputil.Error(w, http.StatusBadRequest, "limit must be a positive integer")
			return
		}
		if parsed > MaxUsersLimit {
			parsed = MaxUsersLimit
		}
		filter.Limit = parsed
	}

	if o := r.URL.Query().Get("offset"); o != "" {
		parsed, err := strconv.Atoi(o)
		if err != nil || parsed < 0 {
			httputil.Error(w, http.StatusBadRequest, "offset must be a non-negative integer")
			return
		}
		filter.Offset = parsed
	}

	users, total, err := h.service.ListUsers(r.Context(), filter)
	if err != nil {
		httputil.HandleError(r.Context(), w, err, errorMappings)
		return
	}

	httputil.Success(w, http.StatusOK, map[string]interface{}{
		"users":  users,
		"total":  total,
		"limit":  filter.Limit,
		"offset": filter.Offset,
	})
}

// GetUser handles GET /users/{id}.
func (h *Handler) GetUser(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")

	user, err := h.service.GetUserByID(r.Context(), id)
	if err != nil {
		httputil.HandleError(r.Context(), w, err, errorMappings)
		return
	}

	httputil.Success(w, http.StatusOK, user)
}

// AdminCreateUserRequest represents admin user creation request body.
type AdminCreateUserRequest struct {
	Email     string      `json:"email" validate:"required,email"`
	Password  string      `json:"password" validate:"required,min=8"`
	FirstName string      `json:"first_name"`
	LastName  string      `json:"last_name"`
	Role      domain.Role `json:"role" validate:"required"`
}

// AdminCreateUser handles POST /users.
func (h *Handler) AdminCreateUser(w http.ResponseWriter, r *http.Request) {
	var req AdminCreateUserRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		httputil.Error(w, http.StatusBadRequest, "invalid json")
		return
	}

	if err := h.validator.Struct(req); err != nil {
		httputil.ValidationError(w, err)
		return
	}

	user, err := h.service.AdminCreateUser(r.Context(), AdminCreateUserInput(req))
	if err != nil {
		httputil.HandleError(r.Context(), w, err, errorMappings)
		return
	}

	httputil.Success(w, http.StatusCreated, user)
}

// AdminUpdateUserRequest represents admin user update request body.
type AdminUpdateUserRequest struct {
	Role      *domain.Role `json:"role"`
	IsActive  *bool        `json:"is_active"`
	FirstName *string      `json:"first_name" validate:"omitempty,max=100"`
	LastName  *string      `json:"last_name" validate:"omitempty,max=100"`
}

// AdminUpdateUser handles PATCH /users/{id}.
func (h *Handler) AdminUpdateUser(w http.ResponseWriter, r *http.Request) {
	targetID := chi.URLParam(r, "id")
	adminUserID := httputil.GetUserID(r.Context())

	var req AdminUpdateUserRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		httputil.Error(w, http.StatusBadRequest, "invalid json")
		return
	}

	if err := h.validator.Struct(req); err != nil {
		httputil.ValidationError(w, err)
		return
	}

	user, err := h.service.AdminUpdateUser(r.Context(), adminUserID, AdminUpdateUserInput{
		UserID:    targetID,
		Role:      req.Role,
		IsActive:  req.IsActive,
		FirstName: req.FirstName,
		LastName:  req.LastName,
	})
	if err != nil {
		httputil.HandleError(r.Context(), w, err, errorMappings)
		return
	}

	httputil.Success(w, http.StatusOK, user)
}

// AdminResetPasswordRequest represents admin password reset request body.
type AdminResetPasswordRequest struct {
	NewPassword string `json:"new_password" validate:"required,min=8"`
}

// AdminResetPassword handles POST /users/{id}/reset-password.
func (h *Handler) AdminResetPassword(w http.ResponseWriter, r *http.Request) {
	targetID := chi.URLParam(r, "id")
	adminUserID := httputil.GetUserID(r.Context())

	var req AdminResetPasswordRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		httputil.Error(w, http.StatusBadRequest, "invalid json")
		return
	}

	if err := h.validator.Struct(req); err != nil {
		httputil.ValidationError(w, err)
		return
	}

	err := h.service.AdminResetPassword(r.Context(), adminUserID, AdminResetPasswordInput{
		UserID:      targetID,
		NewPassword: req.NewPassword,
	})
	if err != nil {
		httputil.HandleError(r.Context(), w, err, errorMappings)
		return
	}

	w.WriteHeader(http.StatusNoContent)
}

// setAuthCookies sets access_token, refresh_token, and csrf_token cookies.
func (h *Handler) setAuthCookies(w http.ResponseWriter, tokens *TokenPair) {
	// Access token cookie - available to all paths
	http.SetCookie(w, &http.Cookie{
		Name:     httputil.AccessTokenCookie,
		Value:    tokens.AccessToken,
		Path:     "/",
		Domain:   h.cookieSettings.Domain,
		MaxAge:   int(h.cookieSettings.AccessTokenDuration.Seconds()),
		HttpOnly: true,
		Secure:   h.cookieSettings.Secure,
		SameSite: http.SameSiteLaxMode,
	})

	// Refresh token cookie - only for /api/v1/auth paths
	http.SetCookie(w, &http.Cookie{
		Name:     httputil.RefreshTokenCookie,
		Value:    tokens.RefreshToken,
		Path:     "/api/v1/auth",
		Domain:   h.cookieSettings.Domain,
		MaxAge:   int(h.cookieSettings.RefreshTokenDuration.Seconds()),
		HttpOnly: true,
		Secure:   h.cookieSettings.Secure,
		SameSite: http.SameSiteStrictMode,
	})

	// CSRF token cookie - readable by JavaScript
	csrfToken := generateCSRFToken()
	http.SetCookie(w, &http.Cookie{
		Name:     httputil.CSRFTokenCookie,
		Value:    csrfToken,
		Path:     "/",
		Domain:   h.cookieSettings.Domain,
		MaxAge:   int(h.cookieSettings.AccessTokenDuration.Seconds()),
		HttpOnly: false, // Must be readable by JavaScript
		Secure:   h.cookieSettings.Secure,
		SameSite: http.SameSiteLaxMode,
	})
}

// clearAuthCookies removes all auth cookies by setting Max-Age=0.
func (h *Handler) clearAuthCookies(w http.ResponseWriter) {
	http.SetCookie(w, &http.Cookie{
		Name:     httputil.AccessTokenCookie,
		Value:    "",
		Path:     "/",
		Domain:   h.cookieSettings.Domain,
		MaxAge:   -1,
		HttpOnly: true,
		Secure:   h.cookieSettings.Secure,
		SameSite: http.SameSiteLaxMode,
	})

	http.SetCookie(w, &http.Cookie{
		Name:     httputil.RefreshTokenCookie,
		Value:    "",
		Path:     "/api/v1/auth",
		Domain:   h.cookieSettings.Domain,
		MaxAge:   -1,
		HttpOnly: true,
		Secure:   h.cookieSettings.Secure,
		SameSite: http.SameSiteStrictMode,
	})

	http.SetCookie(w, &http.Cookie{
		Name:     httputil.CSRFTokenCookie,
		Value:    "",
		Path:     "/",
		Domain:   h.cookieSettings.Domain,
		MaxAge:   -1,
		HttpOnly: false,
		Secure:   h.cookieSettings.Secure,
		SameSite: http.SameSiteLaxMode,
	})
}

// getRefreshTokenFromRequest extracts refresh token from cookie or request body (for backward compatibility).
func (h *Handler) getRefreshTokenFromRequest(r *http.Request) string {
	// Try cookie first
	if cookie, err := r.Cookie(httputil.RefreshTokenCookie); err == nil && cookie.Value != "" {
		return cookie.Value
	}

	// Fallback to request body for API clients
	var body struct {
		RefreshToken string `json:"refresh_token"`
	}
	if err := json.NewDecoder(r.Body).Decode(&body); err == nil && body.RefreshToken != "" {
		return body.RefreshToken
	}

	return ""
}

// generateCSRFToken generates a random CSRF token.
func generateCSRFToken() string {
	b := make([]byte, 32)
	if _, err := rand.Read(b); err != nil {
		// Fallback to less secure but functional token
		return hex.EncodeToString([]byte(time.Now().String()))
	}
	return hex.EncodeToString(b)
}